WIP: Compression Ratio Fix
This is a proposal for how to fix the compression ratio calculation. Previously, chat_template had a surprisingly large impact on the calculated compression ratios. Under this fix, we additionally would run the concat formatter on the prompt in order to get a special-token-free version of the text. The downside, is that we also need to tokenize that, which complicates implementation.

Currently, this is only done for HFLLM. If we agree that this is the way forward, it would be relatively easy to do for other models as well.