Modern dependency structure
The current project specifies dependencies in a poetry-specific way. This PR should not change any dependencies, but uses the standard Python features to achieve the same goal. This will make it easier to switch to different tools in the future.

It also makes it clear, that we're hard-coding some dependencies. We should revisit whether this is actually needed. There are also some typing-specific dependencies that should really be dev-dependencies, but I've left them as is for now.