Modern dependency specification (#15)
The current project specifies dependencies in a poetry-specific way.
This PR should not change any dependencies, but uses the standard Python
features to achieve the same goal. This will make it easier to switch to
`uv` in the follow-up PR:
https://github.com/Aleph-Alpha-Research/eval-framework/pull/16

It also makes it clear, that we're hard-coding some dependencies. We
should revisit whether this is actually needed. There are also some
typing-specific dependencies that should really be dev-dependencies, but
I've left them as is for now.