Manual Package Release workflow
This PR removes the auto-version bumping as part of a move towards a more usual release-workflow through Github. That is, releases are manually triggered through Github releases as documented in the `CONTRIBUTING.md` file.

Going forward, releases through Github will also automatically deploy the package to the respective package indeces.