Cleanup CI (#23)
This PR cleans up the CI pipeline and fixes issues. Next to cleanup, the
main change is to remove extra package permissions and to combine the
two docker image GPU tests: Pulling an image takes up to 10min, so that
cuts down the time our CI pipeline takes by the same amount.