Lazy task registration (#25)
Currently, `eval_framework` imports all tasks before they can be
registered. Importing all of them is not only slow, but also means that
all extra dependencies needed for the task must be installed, which be
annoying for complex dependencies
https://github.com/Aleph-Alpha-Research/eval-framework-companion/issues/127

This PR removes the global task enum and replaces it by a global
registry. Next to the previous functionality, this registry allows lazy
registration and automatically imports the relevant tasks when accessed.

This should allow us to be more fine-grained about the kind of tasks
that we want to register with the framework.