Optional model dependencies (#32)
Right now we have separate extra dependencies for different models, but
we cannot actually use the model file without installing all of them.

This PR splits up the model file and only makes models available if the
corresponding extra is installed. A more permanent solution will require
splitting up the tests too, but I'll leave this for future work.
Together with #27 this will allow us to kick `torch` out of the main
dependencies.