Avoid redis cache load on import time
Decorators replace the decorated function at import time. This meant, that the previous code calls `client = _get_redis_cache()` at import time, rather than on the first function call. This PR aims to change that and call it only when needed.

The caveat is that I am not super familiar with async, so no idea if I broke something. Would be good if someone with more knowledge could double-check.