Fix import_models from eval_framework
Some of our scripts use `import_models` on installed paths. While this is never a good idea, we should support these use-cases that worked in the past. The fix here is to have `import_models` recognize that it is importing from an installed `eval_framework` path and then use the normal import workflows instead of file-imports.