Remove torch dependency (#61)
Now that the blocking [uv
bug](https://github.com/astral-sh/uv/issues/15661) is fixed in the
0.8.16 release of `uv`, we can finally drop `torch` as a main dependency
of `eval-framework` and unblock repos like `scaling` from using it.

I've also removed the cpu/cuda groups, and instead opt for cuda on Linux
machines, and the default pypi channel for anything else. The reason for
that is that, unfortunately, group-specific sources are [still a big
buggy](https://github.com/astral-sh/uv/issues/15730) in `uv`.

Related issue:
-
https://github.com/Aleph-Alpha-Research/eval-framework-companion/issues/150