Reactivate sphyr formatter test (#88)
<!--
For Work In Progress Pull Requests, please use the Draft PR feature,
see https://github.blog/2019-02-14-introducing-draft-pull-requests/ for
further details.

     For a timely review/response, please avoid force-pushing additional
     commits if your PR already received reviews or comments.

Before submitting a Pull Request, please ensure you've done the
following:
     - 👷‍♀️ Create small PRs. In most cases this will be possible.
     - ✅ Provide tests for your changes.
     - 📝 Use descriptive commit messages.
- 📗 Update any related documentation and include any relevant
screenshots.
     - 📗 Reflect the changes you made in the changelog.

-->


## PR Checklist
- [x] Use descriptive commit messages.
- [x] Provide tests for your changes.
- [ ] Update any related documentation and include any relevant
screenshots.
- [x] Check if changes need to be made to docs (README or any guides in
`/docs/`).
- [ ] Reflect the changes you made in the changelog.

## What type of PR is this? (check all applicable)

- [x] Refactor
- [ ] Feature
- [x] Bug Fix
- [ ] Optimization
- [ ] Documentation Update

## Description

## Related Tickets & Documents

<!--
For pull requests that relate or close an issue, please include them
below. We like to follow [Github's guidance on linking issues to pull
requests](https://docs.github.com/en/issues/tracking-your-work-with-issues/linking-a-pull-request-to-an-issue).

For example having the text: "closes #1234" would connect the current
pull
request to issue 1234.  And when we merge the pull request, Github will
automatically close the issue.
-->

- Related Issue #
- Closes #

## QA Instructions, Screenshots, Recordings

_Please replace this line with instructions on how to test your changes,
a note
on the hardware and config this has been tested on, as well as any
relevant
additional information._


## Added/updated tests?
- [x] Yes
- [ ] No, and this is why: _please replace this line with details on why
tests
      have not been included_
- [ ] I need help with writing tests

## [optional] Are there any post deployment tasks we need to perform?

---------

Co-authored-by: philipp.siedler <philipp.siedler@aleph-alpha-ip.ai>