Maximum generated tokens for the same benchmark varies per tokenizer fertility (#105)
## PR Checklist
- [X] Use descriptive commit messages.
- [X] Provide tests for your changes.
- [X] Update any related documentation and include any relevant
screenshots.
- [X] Check if changes need to be made to docs (README or any guides in
`/docs/`).
- [X] Reflect the changes you made in the changelog.

## What type of PR is this? (check all applicable)

- [ ] Refactor
- [X] Feature
- [X] Bug Fix
- [ ] Optimization
- [ ] Documentation Update

## Description
Add a variable to scale the maximum amount of tokens allowed to generate
if this limit is specified by the task. As different models have
different bytes-per-token ratios, this variable belongs in the model
class. It can be overridden in the LLM ARGS in the CLI or in a new class
instance.

## Added/updated tests?
- [X] Yes
- [ ] No, and this is why: _please replace this line with details on why
tests
      have not been included_
- [ ] I need help with writing tests